backend/requirements.txt
boto3
python-jose[cryptography]
requests
passlib[bcrypt]
aws-xray-sdk
